More realistic cycle timings
Fixed more of the opcode timings. All of the cycle counts were
off-by-one, so these great "inc P1" instructions were executing
in 2 cycles instead of 3. D'oh.

This should fix the cycle counts (hopefully), so the performance
metrics are actually realistic.

These give not-so-impressive performance results at the moment,
but there's a ton of room for optimization and I've been squeezing
it to see what I can get out.