Nethub: Allow a backlog of pending transmit messages
Instead of disabling socket recv() when we're waiting on an ACK,
keep reading new packets and buffering them. This fixes a deadlock
where two clients are both waiting on an ACK from each other, but
there's a message packet ahead of the ACK in the kernel's socket
buffer. This change effectively queues all messages, but allows
ACKs to be processed out-of-order.