Sleep after receiving a NAK
If we're talking to a cube whose emulator isn't running,
the nethub won't have anyone to rate-limit our outgoing
packets. So, the client code would spin eating 100% CPU
any time the emulator was not running. This fixes that
by adding a delay after any NAK, effectively having us
slowly poll for the emulator to come back.