Separate TilePool::add() and optimize()
This change makes the act of filling a tile pool and the act of
optimizing it into two totally separate acts- in part for cleanup,
but mostly so that we can do palette optimization fully before starting
tile optimization. This realy helps on the penguin-bg.png sample, to
avoid having literally duplicated tiles in the output. (These tiles
had subtle contours prior to color optimization, but afterward they
were largely the same.)