Hacky loadstream sender
This isn't designed to properly fit within the SDK's design, but to
get some early idea of protocol performance I hacked a quick loadstream
sender into the RadioManager. It is extremely basic, and it doesn't even
send a reset command first. But it's already performing well enough that
the bottleneck is now on the 8051 side, so back over there I go.