Profiler: Keep track of flash idle transitions
To help us optimize the flash decoder, keep track
of the instructions where our flash memory first became
idle. When we're decoding a loadstream, we aim to keep the flash
busy as close to 100% of the time as possible. If the CPU is
lagging behind the flash, it's an opportunity for optimization.