Big SDK reorganization for master firmware
I was getting frustrated by the rather ad-hoc idea of what parts of the 'SDK' were
actually going to be in the real SDK.. so this change rips off that particular band-aid,
and gives us a clear boundary between SDK and master firmware. And within the master firmware,
we're keeping a boundary between code that should be applicable to the real master cube, vs.
code that's only for simulation use.