Implement background asset streaming, mostly.
All of the SDK and radio comms seem to be working nicely...
but there's still something broken. The first few tiles decompress
nicely, then the decoder gets out of sync. I'm not sure if it's
a communications problem (seems unlikely?) or if I introduced a bug
in STIR or flash_decoder with some of the recent refactoring.