Another new codec, DNibRF
This one is inspired by the DVarGeneral, itself a generalization
for statistically modelling DVarNib_s4. This one has some important
performance differences that should make it better suited for use
over the radio. In fact, this encoding is designed to form the outermost
layer of the stream encoding that we send over the radio.