More video mode goodness
Start adhering to the new VRAM layout. This change includes
some dispatch code, the general scaffolding for our video mode code,
and changes to the 'hello' example to keep it compatible with this
new firmware.

This currently implements only a few video modes: a BG0 mode, same
as before, plus new solid-color and 32x32 chunky-pixel modes.