Master side radio implementation
The radio implementation (transmit, ack, timeout) in a fairly
basic form is complete on the master side. It's transmitting packets,
which the buddy cube seems to occasionally receive... but from the
master's POV, everything times out. Time to debug the buddy cube
to see what this looks like for it.