WIP: VRAM access cleanup, and frame flow control
This needs some debugging! Does not work currently, and doesn't compile on ARM yet.
But it seemed more useful to check it in anyway, since it's a big interface change..

This implements _SYS_paint() such that it should actually do what it's supposed
to, rather than being a total stub. It adds system calls to do common VRAM ops,
and implements the userspace VideoBuffer in terms of that.

The new flow control code needs a system clock on the master, so this also implements
a basic notion of a monotonic system clock. But this isn't implemented yet on ARM,
since I need to set up the Cortex-M3's clock, including an overflow interrupt.