BG2: RLE optimization
This assumes that the X axis is major, and watches for runs of identical
values after performing the X accumulator update. If we see a run, we
short-circuit the bulk of the pixel loop to clock out an identical pixel.

To make the X axis major, the master-side code will swap the LCD axes if
necessary.