Timer-based kickoff for I2C accelerometer reads
The radio ISR was way too high-frequency, and it often
came at inconvenient times. Instead, do the thing we actually
wanted, and set up a lower-frequency timer ISR for triggering
accelerometer reads. In this case, at 162 Hz. This also fixes
Timer0/1 in the simulator to run at the right rate for the Nordic
chip.