Higher interrupt priority for I2C
There isn't really any receive buffering to speak of,
and there's no flow control. So we're really on the hook
to get our I2C byte before it's too late. Of course, I2C
is a quite slow bus, so normally that isn't a problem- but
we need our ISR to be higher priority than RF, which can
be a very long-running ISR.