Initial support for other LCD models
This is mostly a structural change to lcd_model.h, so that we
can support both the Giantplus and Truly LCDs. Also changed the
table-driven init so that we can have commands with no data byte
associated.

The initialization table for Truly doesn't seem to work yet. It
might be some other hardware problem, or perhaps the table is
wrong. Waiting to hear back from Truly on this, hopefully.