Big code audit
I've been seeing some very suspicious errors on the master side
when running on hardware. The most suspicious of these was some
packet corruption that looked like unrelated bits getting OR'ed into
the last byte of the radio packet. Ouch! These are proving very
difficult to track down in hardware, so I'm taking another approach:
audit all of the code so far, to look for potential problems, and
add plenty of assertions that are compiled in only on simulator builds.

This patch itself actually catches a few bugs. They'll be fixed
in subsequent patches.