Another, even more heavyhanded, way to undeadlock I2C
I won't even try to explain it, but... we were still getting into
a state where the High pulse on SCL/SDA had no visible effect on SDA,
which was stuck low. So, we'll try pulsing SCL low after driving
the bus high. This is a very brief pulse, just enough to discharge the
bus capacitance really. So far so good, but I won't call this 'done'
until I don't see the lockup for a few days...