Build as separate files, rather than as a single step
The single-step build was done on purpose, for optimization
purposes. The inter-module optimizations between the hardware
modules and the emulation core were very important. But I'm
about to convert this to C++, and I can use classes with inline
functions to do the same thing, only better. So, tidy up the
build first.

This also gets it all building with -Werror -Wall.