Remove delay when simulation is paused
This isn't actually helpful now that we have a physics simulation and user interaction happening.
It's totally reasonable to expect physics to continue to happen at a normal rate even when the
CPUs are stopped- and indeed this might be useful for manually triggering certain kinds of race
conditions that may happen if cubes are moved very rapidly.