Sampling location wrap-around
Sampling locations are supposed to wrap from the beginning of VRAM
to the end, both for correctness and so that we aren't accessing
other (special-purpose) memory in the nRF24LE1. The XDATA exception
caught this, and the fix is just a simple anl instruction.

This looks like it fixes the corruption that would sometimes
show at the top of the 'monsters' demo.