Simulate nRF24LE1 radio retries
This fixes the RX FIFO overflow exceptions. These were real, in that
the simulator was indeed experiencing a FIFO overflow. But real hardware
would NAK those packets, and the sender would retry automatically. I
didn't want to rely on this behaviour in the simulator, but it turns out
that in the remaining cases where we were hitting this, there's a perfectly
legit reason to be taking so long. In most cases, it's a heavily compressed
packet that clears the screen, or does some other repetitive task where
the actual cycle count required to clear all of that memory is greater
than the minimum spacing between packets.