Support setting cube RF address at compile time
This is how we can test with multiple cubes on real hardware for now.
Compile a different firmware for each cube, with CUBE_ADDR defined
(CFLAGS environment var) to an ID number for each one.

This mimicks the default-override scheme used by siftulator's radio
emulation, but notably that emulation does NOT work if you define
CUBE_ADDR, since it only sets the default address that's hardcoded
into the radio. If the cube is overriding it with this hack, it
will do so identically on every simulated cube.