Rough outline for a PID/CRC duplicate check
This is to detect and simulate the condition Liam brought up,
in which the nRF24 can falsely detect a packet as a dupe if it
happens that the PID matches the last PID that the particular cube
has seen, and there's simultaneously a CRC collision or identical
packet payload.

This isn't related to our current bugs probably, since we haven't
tried simulating 4 cubes. But it's something we do want to have a
good fix for in the future, and we need the simulator to accurately
mimic this hardware behaviour.