Fix radio packet loss!!!
Woo... this one turned out to be something ridiculously simple
of course, but it was a lot of tool assistance (tc-siftulator
and my trusty mixed-signal oscope) that finally led to the answer.

There are occasions when the cube processes packets slower than
they arrive over the radio. It's uncommon, but typically the
dense RLE packets that clear the screen or draw simple patterns
will trigger this at least once.

This is normally not a big problem. I was trying to catch them
for optimization purposes, so the simulator was intentionally
flagging this as an exception. But in reality, the ShockBurst
protocol provides enough flow control via its retry mechanism,
and all is well.

Well, almost all. At some point, the cube has to deal with the
fact that it's getting more packets than IRQs. And more specifically,
it needs to not inadvertently drop packets entirely by letting a
receive slip through the cracks when one occurs while we're already
inside the ISR. This particular race condition required a bit more
timing variability to hit, and it showed up very rarely in the
emulator.

The fix is very small and efficient. We can use the implicit STATUS
read that comes with the IRQ acknowledge, at the very end of the
ISR, to check whether we need to loop back up and snag another packet
from the FIFO.