Swap out rand() with rand_r()
Internally, rand() uses malloc() to get some thread-local memory to
put the seed in. Well, we're still undecided on how dynamic memory
allocation is going to be implemented in the master cube.. it's likely
we'll have it in the runtime, but not in the firmware proper. For now,
it's off limits. (malloc() will cause an intentional runtime hang)