Another RX FIFO fix for the radio ISR
This one hasn't shown any symptoms in hardware, but in the simulator I was getting
xruns when running the 'monsters' demo. The trace revealed that we were getting
spurious IRQs, which weren't getting detected at all. So we'd underrun the RX FIFO.

This patch explains why we might sometimes see spurious IRQs, and detects them cheaply.