Neighbor protocol, physical layer
We can receive and transmit 16-bit packets now! There isn't anything
real to send, nor anywhere real to send it yet.. but the individual pieces
seem to work. Tested bit detection on real hardware with the oscope, and
tested the actual received word using DEBUG_REG on the simulator. So far so good.