tc-siftulator: Movable walls
Earlier, I punted on having the world boundaries exactly match the
window size, since I was worried about moving the walls when the aspect
ratio changes. It turns out that this isn't really a problem- so now
we have a perfect match between the window boundary and the physics
boundary.