Windows build cleanups
This is a work-in-progress.. I've been trying to get us to build
out-of-the-box on Windows with just MSYS and Mingw32 gcc. I want
SDL and pdcurses linked statically, so the resulting binary is
easy to use everywhere. And I want the debugger and the command
line I/O to actually work :)

This gets pdcurses building (same as Box2D) so that it's compiled
statically, and we don't need anything external at either build or
runtime

Also switch it to run in console mode, and work around sdl_main.

I was trying to test this in a VM. Graphics doesn't actually work
yet (not sure if it's my bug or VMware's) but I at least fixed the
crashes that were showing up from not having GLSL support.

This patch updates SDL to the latest mingw32 1.2.x release.
I tried to build it statically, but this distribution doesn't
appear to actually have a real static lib. I'm tempted not to spend
much more time on it, since I suspect I'll want to ditch SDL soon
anyway.