Replace SDL with GLFW
SDL is a piece of crap. The "stable" version 1.2 is ancient,
it's LGPL licensed, and it's really not a great fit for what we
need. Our needs are a lot closer to GLUT... but glut sucks too.

So let's try GLFW. It looks like a reimagined GLUT that doesn't
totally take over your app's main loop.