Lazy interrupt dispatch, basic block preemption
This fixes the neighbor sensors in our binary-translated
firmware. The problem: the neighbor sensor is very sensitive
to interrupt latency. There's a reason it runs at super-high
priority. But with SBT, our interrupt latency could be as long
as the longest basic block.

This change allows preempting a basic block and restarting
its timer after the IRQ returns. We add a tick delay to the
IRQ stack, and allow preemption on any clock tick.

This means we need to handle interrupts more frequently.
Since the IRQ code is really too slow to run every clock
tick, this change also implements lazy IRQ dispatch checks.
We only do the dispatch check if a peripheral is firing
an interrupt, we're returning from an interrupt, or a relevant
SFR is modified.

Tested neighbor.sim with SBT, interpreter, and debug mode.