Fix one of the I2C bugs
This one was a bug in the simulation- we weren't resetting the
hardware I2C state machine when the peripheral is disabled then
reenabled.

The other bug is actually in the firmware- it's a bug in the IRQ
handlers that can occur when we reset I2C while a lower-priority
I2C interrupt is already in progress. I'll fix that in a separate checkin.