Don't reinitialize pdcurses when adding/removing cubes
This didn't work properly on Windows, and on Mac it worked
but it was still really annoying.

We have to stop and restart our emulation thread in order to
change the number of cubes. Previously this involved tearing
down curses too, since we set up curses on that thread.

This change moves the curses setup/teardown outside that thread,
and keeps it up even across thread restarts.