Don't poll for flash data per-scanline
This was kind of a bad idea. I originally implemented
this to help improve flash utilization, but that was when
we were always rendering continuously. Now that we have
better frame rate control, and we always want graphics to
be higher priority than flash.. this is seeming kind of
counter to our goals.

This change may allow some further optimization to the
graphics code, since it no longer has to be written to
be interruptible at any time by flash.