EXE compression
There are a lot of large, repetetive data tables in
the emulator. For example, we have a 128 kB lookup
table that's used to locate binary translated basic blocks.
This table is highly repetetive, and it would easily compress
down to a tiny fraction of its original size. We also have
the data ROM, which is mostly zeroes (representing code that
was removed).

Instead of adding a bunch of one-off compression schemes to
our code, it's easier to just pack the whole binary. I tried
UPX, and it has no noticeable impact on startup times.