Fixed flash scheduling, and flash deadlock
The scheduling fix from yesterday was mostly working,
but it was triggering a separate bug that could cause
the flash programming code to deadlock. That turned out
to be a very legitimate race condition- flash_program_start()
assumes that the last programming operation has already finished.
If that isn't the case, it uses the wrong Data# state, and any
subsequent waits will end up with their condition inverted.
Hence, infinite loop.