New, updated scroller example for BG0+BG1
This is a much more complex scrolling demo. So much so
actually, that it fails dismally on current firmware, due
to several bugs in the BG1 code. So, think of this as test-driven
development. First, we check in a failing test case.

This has an endlessly-scrolling BG0 layer, and a BG1 layer with both
a static overlay and a smoothly animating sprite.