Support for text files in bin2c
By explicitly specifying the mode as text, we'll do CRLF conversion
appropriate to the current platform. This means we'll always have
LF newlines in the generated data arrays, which is what we want.

I noticed this when building on Linux, and seeing that data.cpp was
regenerated with a new set of line endings :)