Modify MDU emulation and MDU usage to reflect reality
When I wrote this, I was just guessing as to how the MDU
really works in this somewhat-undocumented usage. I wasn't
far off- turns out the real hardware really does require writes
to both MD0 and ARCON in order to trigger a shift. Update the
emulator to account for this, and update the BG1 code.