New chroma key representation
Instead of defining a single color as a chroma key, while casually
brushing past the fact that we only ever check the MSB.. let's define
the colorkey *only* via the MSB, and use the LSB to encode additional
data that may help the firmware.

The first such piece of data is an "End Of Line" flag, that tells the
firmware that all remaining pixels on the tile's scanline are transparent.