More aggressively declassify BG1 lines
It's very much to our advantage to choose a non-BG1
scanline renderer whenever possible. One common case that
we missed is when a line has BG1 tiles on it, but due to
horizontal scrolling they are fully to the left or right
of the screen. This change adds a check for that. We can
detect this condition by looking at the low 17 bits of the
MDU, but to do this we need to rearrange the setup sequence
a bit so that the result of the MDU shift is available when
we need it.

All tests pass.