RAISING THE BAR on fucking insane scanline renderers
This is a fast renderer for BG0 plus a single sprite.
It segments the scanline in upto seven fragments, in
order to handle tile-size bursts of BG0 both before and
after the sprite data. The sprite is still colorkeyed
on a per-pixel basis, but with less generality and a bit
more speed than the general-purpose sprite renderer.