Style and build fixes.
SPACES, guys.

And redundant naming sucks. Let's keep all the names
semantically meaningful. The changes to SYSEventVectors
were using inconsistent terminology- are they vectors
or are they callbacks? Pick one, not both.

Also, the recent changes to ABI broke the cube firmware
build. This file needs to be valid C and C++ code, and it
needs to build with GCC and SDCC.