Don't resend reset tokens upon receiving first ACK
This was a really bad idea; we can end up sending two resets in
quick succession, and if we're unlucky we'll start an asset
download between the two corresponding ACKs. This means we could
reset partway into the asset download (decoder state corruption)
or we could end up with an off-by-one FIFO size count (FIFO overflow
OR a rollover and deadlock).

This was behind most of the asset load failures we've been seeing.

The rest of them seem to be related to cases where we start
an asset download when the ACK state is unknown. We dont' have
an explicit way to find out the baseline values for ACK packets...
we may want to just explicitly include this in our cube pairing process later.