Fix cube firmware hang when running in simulator
This hang was occurring in the new accelerometer init code. It runs early
in startup, and it happened to require the I2C IRQ to at least be routed
correctly (so the IRQ flag could be observed) even if the IRQ handler wasn't
intended to be invoked. This meant that, at the time the accelerometer init
code runs:

 - The IEX3 interrupt was still falling-edge triggered. We wanted rising edge.
 - The "Data Ready" bit in W2CON1 would have been set. This only occurs in the
   simulator; real hardware does not implement writing to this bit.

These two bugs each prevented the interrupt flag from being properly set when
the I2C transfer completes, leading to an infinite loop in i2c_tx()