Neighbor side cleanups
Make the relationship between side and binary value more clear in the
firmware, and use this same consistent enumeration throughout the entire
stack. The hardware change from the prototype to rev 1, then, is implemented
by changing the pin definitions in the simulator and the firmware.

Tested the 'sensors' demo, all is well.