Mask off unneeded sides when receiving neighbor payload bits
After receiving the two-bit side code, we can use that information
to mask off all receivers other than the one we're actually using.

In a perfect electrical environment (like the simulator) this change
makes no difference; but in physical hardware, it serves as a check
for the side bits. Without this check, corrupted side bits would
be interpreted as valid.

Tested on the simulator with the 'sensors' demo. Not yet tested on
real hardware.