Maybe fix mixing bug
I'm having trouble testing this on my laptop, since it's a bit slow to keep
up with running games full-speed in simulation anyway... but it seems like this
might fix "volume changes erratically":

https://www.pivotaltracker.com/story/show/22913217

Also, my previous comments about the mixing were wrong; I misread the code, and
the MAX operation I saw was actually in reference to the length of a mixing buffer,
not the actual sample data. I removed the bogus comments, and to clarify the actual
operation of this code I renamed the channel's pullAudio() to mixAudio().

The bug seemed to be that i mixAudio(), the original buffer contents would be added
twice. This means that, if two samples play simultaneously, the first sample will
play at twice the volume. That would seem to explain the bug!