Implement emulation support for CE
Previously, we had no real support for the CE (radio enable) input.
We'd store it, but it was still possible to enqueue packets when the
radio was disabled.

This change makes the emulation more accurate. Packets received with
CE=0 are dropped without acknowledgment, and we log this condition to
the trace log.