Move radio_rx_enable(), and explain why
We want to only enable the radio hardware after the system is fully set up, and
the global interrupt enable has been set. This is partly because it just seems like
a good practice anyway, keeping the power-hungry radio off until we're ready for it.
But more specifically, it can fix a bug where if a packet is received during init,
its interrupt flag will get discarded by some initialization code (in I2C) which does
not preserve the value of IRCON.