Discard stale ACK packets received from newly connected cubes
We assume that the first ACK packet received is 'stale', because it's
been sitting in the cube's ACK buffer for an unknown length of time.
It may reflect an old status, from before another operation (like a
flash write that a previous incarnation of the master firmware started).

To make sure we have a freshly-assembled ACK packet to use as the baseline
for all our deltas, we wait until the second ACK. We're guaranteed to receive
this ACK eventually, due to the flash reset state machine.