Break basic blocks after a radio SPI write
This may fix false SPI undeflow exceptions, which seem to be caused
by inaccurate timing simulation when we're running in SBT mode. In
particular, the exceptions I'm seeing tend to occur in the one small
loop in our Radio ISR which buffers asset data. Without this patch,
that loop would translate to a single basic block, and the simulator
would have no opportunity to advance time within that basic block.

Testing: Asset loading for the 'word' example, no hangs so far.