Simulator performance: Check T1/2/3 edges only when necessary
Right now we do the timer input edge check on every clock tick,
but really it's rarely necessary; right now we only use this feature
for neighbor inputs.

So, rearrange the timer code to do this check only on the slow path,
and add an explicit flag to trigger the check so that we can do so
when a neighbor pulse is propagated.