Explicit SPI wait during flash-escape copy
This one stupid loop, responsible for copying flash-escaped data
to the programming FIFO, has been endlessly troublesome. Even though
on paper it should work fine, the combination of SPI and a single-basic-block
loop keeps causing rare and frustrating SPI underrun exceptions in simulation.

I still don't have any reason to believe that these underruns are occurring
on real hardware, but really this loop is pretty pathological, and I'm okay
with making it a bit more conservative in order to allow us to run faster
in simulation. We can gain a lot of speed if we decouple the hardware timing
from the CPU timing at least up to the basic-block level. That's what SBT tries
to do, and it's why we keep hitting these underruns when running via SBT.

Testing: gfxdemo loads properly, even with my aggressive performance patches for SBT.