Build fix for Mac OS
This fixes a fatal warning,
"format not a string literal and no format arguments".

It was never an actual problem, since this function _is_ expecting
a format string, which just happens to have zero arguments. But we
can make GCC happy without any real trouble just by dropping in a
token "%s" and using the degenerate format string as a plain string
argument.