Very crude fix for radio power-on delay
Liam's hunch was right; the radio problems we were seeing on reset
were due to us trying to transmit too soon after power is applied
to the radio.

This patch adds a synchronous delay hack based on a dummy SPI transfer.
It also includes a comment that explains what's going on, and suggests
a better implementation.