Disable STRIP_NB_MODE, it doesn't work :(
Unfortunately, the way speex is designed, wide-band
encoding builds on the encoding used for narrow-band.
In fact, during the wideband decode process, we do a
full narrow-band decode on part of the incoming packet.

So, it isn't really possible to strip out any nontrivial
amount of code from the narrow-band codec. Disable
this flag.