More comprehensive pixel zoom detection
Now we detect zoom levels that are any small integer multiple,
not just 1:1. Keyboard shortcuts now exist for 1x and 2x zoom.

Additionally, we are now setting the filtering mode dynamically
for each cube, based on what is needed. Anisotropic filtering is
only enabled when cubes are tilted, and in the pixel-accurate case
we disable filtering altogether. (This is required for 2x and higher
zooms, since the polynomial filter starts to affect the pixel edges
in those cases.)