Simulator bugfix, black LCD screens after full-screen
When transitioning between full-screen and windowed modes,
we lose our OpenGL context and must re-init it. However,
this needs to happen entirely within the GL renderer, so
the renderer needs access to the LCD framebuffer even if
the framebuffer hasn't changed.

This change accomplishes that by using a separate flag to
indicate if the LCD contents have changed.