Disable filtering in our fragment shader in pixel-accurate mode
This isn't so much necessary for correctness, since we're already
disabling filtering in our sampler, but this does let us turn off
antialiasing for the edge of the LCD, so that we don't have this
annoying single blended pixel across the perimeter of the
otherwise totally accurate LCD.

Since we now compile the shader twice, with an #ifdef for filtering,
this means that the filtering code is completely disabled when it
isn't being used. This should help reduce our GPU load even further.