Snap to right angles during rendering
If we're really close to a right-angle rotation, snap to it exactly
and avoid breaking pixel accuracy.

It's really easy to get cubes close but not quite at a right angle,
for example if you're pushing+aligning one cube, but bumping that into
a neighboring cube.

With this change, these small physical interactions will still look
realistic, but if the cubes still look straight they'll actually
render that way, and we can stay on the pixel-accurate rendering path.