Fullscreen to desktop resolution
The hardcoded resolution sucks. Instead, pick the desktop resolution,
so that we don't actually change video modes. This is usually what we
want, so we can maintain pixel accuracy.

The only downside, is that on a slow GPU this may be too many pixels
to push, and we may slow down. That problem should be fixed by optimizing
our graphics rendering :)