Fix lightmap banding
The 8-bit light map had visible banding, which looked especially
bad in full-screen. This was purely due to the limited resolution
of the image file.

Part of the problem was that the original lightmap only used a
small portion of the dynamic range available to it. This change
modifies the lightmap to use the full 8-bit dynamic range, and it
adds ambient lighting as a separate term in the fragment shader.