Re-upload LCD texture when the filtering mode changes
This fix is necessary to avoid a black screen (blank texture) in cases
where we need a mipmap but the last upload was made with mipmap generation
disabled. Now we always upload a texture if the mode changes.

This change also lets us only change the filtering state when we're uploading
a new frame, which may improve performance if we were causing OpenGL state thrashing.