Use separate GL textures for filtered vs. pixel-accurate modes
This change is primarily a workaround for a specific GPU bug
that Eric noticed. His GPU drivers seem to be failing to enable
mipmap generation dynamically in some cases. To work around this,
and possibly to increase performance by reducing OpenGL state thrashing,
this patch switches to two separate texture pools. Each pool has been
statically configured with the right mipgen and filtering options
for one of the two rendering modes.