LLVM 3.0 binaries for 64-bit Mac OS
Compiled myself on Lion, based on the released LLVM 3.0
and Clang 3.0 sources. Enabled ARM and Host targets, and
stripped down the resulting binaries and libraries to what
you see here.

(Most notably, removed binaries that didn't seem useful to us,
and removed the libraries for all code generators on the assumption
that we won't be programmatically generating code for any physical
CPU architectures.)