Ineffectively frob some LLVM flags, to make lli run better
This is really just doing a bunch of the standard compile-time
optimizations at link-time, with some added optimizations that
I hope will lower some of the operations that are unsupported
in LLVM's interpreter.

It doesn't really help much, but it seems to help some. In any case,
this change is pretty short-lived.