Numeric syscall symbols
Instead of embedding a table in the compiler, it seems to make much more
sense to keep that info centralized in abi.h. So, instead, teach the compiler
to allow special numeric _SYS_<n> symbols, and use the __asm__ keyword to alias
those syscall definitions to the proper symbol name.