Wacky decorated symbol approach for call fixups
Kinda unconventional, but it works...

The sanest way to fit our requirements into LLVM seemed to be creating
magic symbol aliases, which then get converted into the various forms
of 'decorated' symbols by the linker. For example, a symbol of the form
_call$Foo will get linked as a CALL operation to the absolute address
of symbol Foo.

The linker still isn't calculating real segment addresses, but this
is otherwise looking very close to correct.