Unconditionally check the poweroff timer
It isn't correct to use the main loop's concept of "idle" to optimize
the poweroff check, since the main loop considers "idle" to mean no
graphics or flash activity at all. However, even with the master
disconnected, the cube still could be rendering to the screen. In this
state, the cube will never go to sleep. Oops.