ELF section/segment/metadata overhaul
This is a big refactoring and partial rewrite of the ELF section
and segment handling. All knowledge of how segments are laid out
in memory and on disk has been encapsulated in a new SVMMemoryLayout
object, the stock LLVM 3.0 support for ELF string and symbol tables
has been rolled into a new SVMELFMetadataBuilder object, and the
SVMELFProgramWriter is left to orchestrate things.

This also adds a "-g" switch, which adds debug information to the
resulting ELF file. Currently this means that with -g you get a full
symbol table and a full section table, whereas without -g there are
no sections, only program headers.