New 32-bit instruction alignment implementation
Instead of inserting nops in SVMAsmPrinter, we'd really like to
do this in an earlier compilation pass. Partly this is so we can
try to fill these slots with useful instructions instead of nops,
and partly it's so that we can do a later pass to break the code up
into flash blocks, with the prior knowledge of how large all MBBs
are.

This implements a really simple but not completely useless align
slot filler. It also fixes up some of the tblgen data so that we're
correctly identifying instructions with no unmodeled side effects.