Lots of infrastructure to support late long-branch insertion
To insert a Long Branch op late in compilation (Say, within
SVMLateFunctionSplitPass) we need to be able to create new constant
pool entries which refer to decorated symbols.

LLVM makes this pretty hoopy, as far as I can tell. The MBB at this
point only has an MCSymbol to identify it, and a constpool entry
requires some form of Constant, which can be a GlobalValue but not
an MCSymbol or MCExpr.

So, this patch adds a target-specific MachineConstantPoolValue
subclass. It should be extensible to do other nifty things, like
storing small rodata items in the constpool to improve cache
locality.

This is almost working; the assmebly looks good, but ELF output
is busted because the temporary symbol is apparently being deleted
or ignored.