ABI break: Change how system vectors work
I originally anticipated that the vector table would be mapped
directly into user memory. The problem with that, in practice, is
that we'd be forced to reserve a concrete chunk of address space
for it in the ABI, and it would be difficult to add new syscalls
later unless we'd already reserved space in this table.

Now that we understand how the game runtime VMs will work, we know
it's just as efficient to use setter/getter syscalls for the vectors.
This also lets us make some nice improvements, like allowing every
syscall to include a void* that can be used for an object instance
pointer or other tasty closure goodness.