Fix BB and MBB label mess
I found the root cause for the MBB label trouble earlier: a false
fall-through was being detected by AsmPrinter, causing the label
to be omitted when actually we really needed it. Also, I wasn't
updating the CFG, which is actually kind of important still.

This fixes those root causes and unhacks a bunch of the MBB/BB
stuff in LateFunctionSplitPass. Fixes label redefinition bugs
in compiling the nyan example.