Cleaner implementation of SPAdj decoration
Minor lapse in reasoning... the LOCAL symbol lookups need to understand
SPAdj too, since any time we take the address of a function, it needs that
value. This includes but is not limited to the ELF entry point address.

So, in hindsight, this is exactly what I was doing before. This change
mostly just unbreaks the original LOCAL SPAdj decoration, and comments
what it's doing so that I don't break it again later.