Progress on SVM memory refactoring
This is mostly centered around changing the caching layer to be a bit
more realistic, and to integrate this new layer in with the rest of the
firmware.

I'm about to rip out the ID-based asset loading, so that I can use
our real flash caching layer and address translation scheme everywhere-
but that will be another big commit.