Basic event dispatch for SVM
Event dispatch works by creating a synthetic call() on our way
out of an svc() handler. Typically a syscall will contain a 'yield point',
where events are allowed to be dispatched. Instead of taking effect
right away, this needs to take effect after all side-effects from the
syscall are complete. This includes assignment of the syscall's
return value to registers.

So, the runtime is now responsible for calling Event::dispatch()
at the proper time at the very end of svc(), and it includes
public functions to craft the synthetic call().

In order to limit the amount of user stack space we use, this
system is restricted to dispatching one event at a time. For most
events this is no problem at all; but for neighbor events it will
require some big changes. That's next.